requirements.txt
chex
jax[cpu]
simpleeval
dm-haiku
setuptools
astropy
numpy
networkx
numpyro
matplotlib